Move variables to commons module
