+++ commons/pom.xml
@@ -3,19 +3,19 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
+    <parent>
+        <groupId>me.u8092</groupId>
+        <artifactId>watchdog</artifactId>
+        <version>0.2.0d</version>
+    </parent>
 
-    <groupId>me.u8092</groupId>
-    <artifactId>watchdog</artifactId>
-    <version>0.2.0d</version>
-    <packaging>pom</packaging>
-    <modules>
-        <module>bukkit</module>
-        <module>commons</module>
-    </modules>
+    <artifactId>watchdog-commons</artifactId>
+    <version>1.0.0</version>
 
     <properties>
         <maven.compiler.source>17</maven.compiler.source>
         <maven.compiler.target>17</maven.compiler.target>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     </properties>
+
 </project>
