+++ commons/src/main/java/me/u8092/watchdog/commons/variables/BooleanVariable.java
@@ -1,7 +1,4 @@
-package me.u8092.watchdog.variables;
-
-import me.u8092.watchdog.Main;
-import me.u8092.watchdog.util.DebugUtil;
+package me.u8092.watchdog.commons.variables;
 
 import java.util.List;
 
@@ -35,10 +32,7 @@ public class BooleanVariable {
     public String getName() { return name; }
     public boolean getValue() { return value; }
     public boolean getDefaultValue() { return defaultValue; }
-    public void setValue(boolean value) {
-        if(Main.getInstance().getConfig().getBoolean("debug")) DebugUtil.info("Set BooleanVariable " + name + " to " + value + " (" + owner + ")");
-        this.value = value;
-    }
+    public void setValue(boolean value) { this.value = value; }
     public List<String> getTrueEvents() { return trueEvents; }
     public List<String> getFalseEvents() { return falseEvents; }
     public List<String> getSwitchEvents() { return switchEvents; }
